refact(Maze Scripts) : class 이름 변경
