+++ Assets/Scripts/Map/Puzzle/Maze/MazeDarkness.cs
@@ -1,62 +1,64 @@
-using System;
 using System.Collections;
-using System.Collections.Generic;
+using playerCharacter;
 using UnityEngine;
 using UnityEngine.Rendering.Universal;
 
-public class MazeDarkness : MonoBehaviour, IEventTriggerable
+namespace Map.Puzzle.Maze
 {
-    private Light2D globalLight;
-    private GameObject playerLight;
-    private GameObject player;
-    private bool isInMaze = false;
-    private const float changeRatio = 0.01f;
-
-    private void Start()
+    public class MazeDarkness : MonoBehaviour, IEventTriggerable
     {
-        player = GameObject.FindGameObjectWithTag("Player");
-        globalLight = GameObject.Find("GlobalLight2D").GetComponent<Light2D>();
-        playerLight = player.transform.Find("PlayerLight").GetComponent<Light2D>().gameObject;
-    }
+        private Light2D globalLight;
+        private GameObject playerLight;
+        private GameObject player;
+        private bool isInMaze = false;
+        private const float changeRatio = 0.01f;
 
-    private void OnTriggerExit2D(Collider2D other) 
-    {
-        if (!ConditionManager.Instance.Conditions.ContainsKey("spring_puzzle1_clear"))
+        private void Start()
         {
-            if (other.CompareTag("Player"))
-            {
-                bool previousState = isInMaze;
-                isInMaze = player.transform.position.x < transform.position.x;
+            player = PlayerCharacter.Instance.gameObject;
+            globalLight = GameObject.Find("GlobalLight2D").GetComponent<Light2D>();
+            playerLight = player.transform.Find("PlayerLight").GetComponent<Light2D>().gameObject;
+        }
 
-                if (previousState != isInMaze)
+        private void OnTriggerExit2D(Collider2D other) 
+        {
+            if (!ConditionManager.Instance.Conditions.ContainsKey("spring_puzzle1_clear"))
+            {
+                if (other.CompareTag("Player"))
                 {
-                    StopAllCoroutines();
-                    StartCoroutine(ChangeIntensity(isInMaze));
+                    bool previousState = isInMaze;
+                    isInMaze = player.transform.position.x < transform.position.x;
+
+                    if (previousState != isInMaze)
+                    {
+                        StopAllCoroutines();
+                        StartCoroutine(ChangeIntensity(isInMaze));
+                    }
                 }
             }
         }
-    }
 
-    public void Trigger()
-    {
-        StartCoroutine(ChangeIntensity(false));
-        isInMaze = false;
-    }
-
-    private IEnumerator ChangeIntensity(bool isInMaze)
-    {
-        float targetIntensity = isInMaze ? 0f : 1f;
-        float Ratio = 0f;
-
-        playerLight.SetActive(isInMaze);
-        while (Mathf.Abs(globalLight.intensity - targetIntensity) > 0.01f)
+        public void Trigger()
         {
-            Ratio += Time.deltaTime;
-            globalLight.intensity = Mathf.Lerp(globalLight.intensity, targetIntensity, Ratio);
-            yield return new WaitForSeconds(0.05f);
+            StartCoroutine(ChangeIntensity(false));
+            isInMaze = false;
         }
-        globalLight.intensity = targetIntensity;
 
-        yield return null;
+        private IEnumerator ChangeIntensity(bool isInMaze)
+        {
+            float targetIntensity = isInMaze ? 0f : 1f;
+            float Ratio = 0f;
+
+            playerLight.SetActive(isInMaze);
+            while (Mathf.Abs(globalLight.intensity - targetIntensity) > 0.01f)
+            {
+                Ratio += Time.deltaTime;
+                globalLight.intensity = Mathf.Lerp(globalLight.intensity, targetIntensity, Ratio);
+                yield return new WaitForSeconds(0.05f);
+            }
+            globalLight.intensity = targetIntensity;
+
+            yield return null;
+        }
     }
 }
